Refactor how hooks are organized
- Create a hooks directory
- Hooks now each have their own file
