+++ plugins/ui/src/deephaven/ui/_internal/shared.py
@@ -1,4 +1,4 @@
-from .render import RenderContext
+from ..render import RenderContext
 from typing import Optional
